+++ main_project/src/components/common/tabs.tsx
@@ -2,97 +2,21 @@ import { Tab, TabList, TabPanel, Tabs } from "react-tabs";
 import "react-tabs/style/react-tabs.css";
 import DiaryHome from "../../pages/DiaryHome";
 import MoodChart from "../../pages/moodChart";
-import { useState, useRef, useEffect, RefObject } from "react";
-import ProfileModal from "./Modal/ProfileModal";
-import { axiosFetcher } from "../../api/axiosFetcher";
-import { useAuthStore } from "../../store/useAuthStore";
+import TopBarContainer from "./TopBarContainer";
 import { useSearch } from "../../hooks/useSearch";
-import SearchBar from "./SearchBar";
-import UserMenu from "./UserMenu";
-import { useNavigate } from "react-router-dom";
-import authApi from "../../api/authApi";
-import CustomConfirmModal from "./Modal/CustomConfirmModal";
 
 function MyTabs() {
-  const [showDropdown, setShowDropdown] = useState(false);
-  const { refreshToken, clearAuth } = useAuthStore.getState();
-  const dropdownRef = useRef<HTMLDivElement | null>(null);
-  const [isProfileOpen, setIsProfileOpen] = useState(false);
-  const [showLogoutModal, setShowLogoutModal] = useState(false);
-  const navigate = useNavigate();
   const {
     showSearch,
     setShowSearch,
     searchQuery,
     setSearchQuery,
     isSearching,
-    searchResults,
     clearSearch,
     handleSearchInputRef,
     handleSearch,
+    searchResults,
   } = useSearch();
-
-  const [modalUser, setModalUser] = useState({
-    nickname: "",
-    profileImage: "",
-    introduction: "",
-    preferredGenres: [] as string[],
-  });
-
-  const handleClickOutside = (event: MouseEvent) => {
-    if (dropdownRef.current && !dropdownRef.current.contains(event.target as Node)) {
-      setShowDropdown(false);
-    }
-  };
-
-  useEffect(() => {
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => document.removeEventListener("mousedown", handleClickOutside);
-  }, []);
-
-  const handleOpenProfile = async () => {
-    try {
-      const res = await axiosFetcher.get<{
-        profile_image: string;
-        member: {
-          nickname: string;
-          introduce: string;
-          favorite_genre: string[];
-        };
-      }>("api/members/profile/");
-      console.log("프로필 API 응답:", res);
-
-      setModalUser({
-        nickname: res.member.nickname ?? "",
-        profileImage: res.profile_image ?? "",
-        introduction: res.member.introduce ?? "",
-        preferredGenres: res.member.favorite_genre ?? [],
-      });
-
-      setIsProfileOpen(true);
-    } catch (error) {
-      console.error("프로필 불러오기 실패:", error);
-      alert("프로필 정보를 불러오는데 실패했어요.");
-    }
-  };
-
-  const handleLogout = async () => {
-    try {
-      if (!refreshToken) {
-        alert("유효하지 않은 토큰입니다. 다시 로그인해주세요.");
-        return;
-      }
-
-      await authApi.logout({ refresh_token: refreshToken });
-
-      clearAuth();
-      alert("로그아웃되었습니다.");
-      navigate("/");
-    } catch (error) {
-      console.error("로그아웃 실패:", error);
-      alert("로그아웃 중 오류가 발생했습니다.");
-    }
-  };
 
   return (
     <Tabs className="bg-[#A6CCF2] min-h-screen flex flex-col" defaultIndex={0}>
@@ -109,24 +33,16 @@ function MyTabs() {
         >
           나의 감정발자취
         </Tab>
-        <div className="flex items-center gap-2 ml-auto">
-          <SearchBar
-            showSearch={showSearch}
-            searchQuery={searchQuery}
-            setSearchQuery={setSearchQuery}
-            handleSearch={handleSearch}
-            clearSearch={clearSearch}
-            setShowSearch={setShowSearch}
-            handleSearchInputRef={handleSearchInputRef}
-            isSearching={isSearching}
-          />
-          <UserMenu
-            showDropdown={showDropdown}
-            setShowDropdown={setShowDropdown}
-            dropdownRef={dropdownRef as RefObject<HTMLDivElement>}
-            handleOpenProfile={handleOpenProfile}
-          />
-        </div>
+        <TopBarContainer
+          showSearch={showSearch}
+          setShowSearch={setShowSearch}
+          searchQuery={searchQuery}
+          setSearchQuery={setSearchQuery}
+          isSearching={isSearching}
+          clearSearch={clearSearch}
+          handleSearchInputRef={handleSearchInputRef}
+          handleSearch={handleSearch}
+        />
       </TabList>
 
       <div className="w-full max-w-[1130px] mx-auto">
@@ -141,23 +57,6 @@ function MyTabs() {
           <MoodChart />
         </TabPanel>
       </div>
-
-      <ProfileModal
-        isOpen={isProfileOpen}
-        onClose={() => setIsProfileOpen(false)}
-        user={modalUser}
-      />
-      <CustomConfirmModal
-        type="logout"
-        title="로그아웃 하시겠습니까?"
-        message="로그아웃하면 다시 로그인해야 해요"
-        isOpen={showLogoutModal}
-        onClose={() => setShowLogoutModal(false)}
-        onConfirm={handleLogout}
-        isDanger
-        confirmText="로그아웃"
-        cancelText="취소"
-      />
     </Tabs>
   );
 }
